backend/michicondrias_adopciones/requirements.txt
fastapi[all]>=0.111.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0

uvicorn[standard]>=0.30.0
httpx>=0.27.0
mangum>=0.17.0
